+++ app/presenters/commitment_presenter.rb
@@ -1,0 +1,5 @@
+class CommitmentPresenter < SimpleDelegator
+  def value
+    ActionController::Base.helpers.number_to_currency(super, unit: "£")
+  end
+end